Terminado a DAO de Tecnologia, incluido a flag de ativo.

+++ src/java/br/unesp/rc/habilidades/dao/CargoDAOImpl.java
@@ -117,7 +117,7 @@ public class CargoDAOImpl implements CargoDAO {
         if (con != null) {
             try {
                 pstm = con.prepareStatement(SELECT);
-                pstm.setLong(1, idCargo);
+                pstm.setInt(1, idCargo);
                 res = pstm.executeQuery();
                 
                  if (res.next()){
